move rex_base_url fixture to webview file
