--- fixtures/rex.py
@@ -1,17 +1,0 @@
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this
-# file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-import pytest
-
-from tests.utils import skip_if_destructive_and_sensitive
-
-
-@pytest.fixture
-def rex_base_url(request):
-    """Return a base URL for REX"""
-    config = request.config
-    base_url = config.getoption("rex_base_url") or config.getini("rex_base_url")
-    if base_url is not None:
-        skip_if_destructive_and_sensitive(request, base_url)
-        return base_url
