+++ fixtures/webview.py
@@ -33,6 +33,16 @@ def webview_base_url(request):
 
 
 @pytest.fixture
+def rex_base_url(request):
+    """Return a base URL for REX used for integration testing"""
+    config = request.config
+    base_url = config.getoption("rex_base_url") or config.getini("rex_base_url")
+    if base_url is not None:
+        skip_if_destructive_and_sensitive(request, base_url)
+        return base_url
+
+
+@pytest.fixture
 def webview_instance(webview_base_url):
     url = tldextract.extract(webview_base_url)
     if url.subdomain:
